refactor(roles): Simplify Fish shell config abbrs
Removed conditional logic for setting abbreviations based on
the availability of the `eza` command.

Now, the abbreviations default to using:

    - `l` -> `ls`
    - `ll` -> `ls -lh`
    - `lla` -> `ls -lah`

This simplifies the configuration and ensures consistent
behavior regardless of the presence of `eza`.

As `eza` is not a mandatory dependency for our fish setup
we remove it's setup for it.

We will create in a feature commit a role for installing
and setting up `eza`. In short, better modularizaton.