feat(roles.setup_tmux): Add tmux session management
Bind the key combination `Ctrl+A, e` for launching the session
manager.

Ensure `fzf` is installed as a dependency for project selection

The session manager has some default path it will search for
project in using `fzf`:

- ~/.config/
- ~/Documents/work/
- ~/Documents/github/

The other way to use the session manager is to give it a path
as a parameter.