feat(roles.setup_nvim): Add default editor options
The Lua files are organized under the "me" namespace instead of
using the actual username, keeping things simple and consistent
.
The Ansible task now copies these configs to the user's home
directory, ensuring the right permissions are set.

Those are the options enabled by default:

- Line numbers: Enable line number and relative line numbers.

- Tabs & Indentation: Set tab width to 4 spaces and use spaces
  instead of tabs.

- Line wrapping: Disable line wrapping for a cleaner view.

- Search settings: Use smart case for searches, making it case
  insensitive unless uppercase is used.

- Cursor line: Highlight the current line for better visibility.

- Line width limit: Set at 80 characters.

- Appearance: Enable true color support and set a dark background.

- Backspace: Allow backspacing over indentation, EOL, and start.

- Clipboard: Use the system clipboard for yanking text.

- Split windows: Open new splits to the right and below.

- Keywords: Include '-' as part of a word for easier editing.

The Lua files are organized under the "me" namespace instead of
using the actual username, keeping things simple and consistent.
The Ansible task now copies these configs to the user's home
directory, ensuring the right permissions are set. It's a solid
start for customizing Neovim!