feat(roles.setup_tmux): Add vim-tmux-navigator
This plugin allows easy navigation between tmux panes and vim splits.

The movement is done using `Ctrl + {H, J, K, L}` without the need to
use the tmux prefix, `Ctrl+A` for this configuration.

This commit contains also a fix for programatically installing the
TMUX plugins. It seems the initial incantation was wrong, we had
to remove the `run-shell` command, but call the `install_plugins.sh`
script directly.

Also we made sure Ansible executes the plugin installation using
`bash`.