feat: Remove `systemd` and use `fish` to initialize `ssh-agent`
Now `ssh-agent` is no longer initialized by a `systemd` service. This
helps us to make `forger` container friendly. Now `ssh-agent` is
initialized by a `fish` plugin script. Basically a `fish` script that
lives in `~/.config/fish/conf.d/` directory.

During the execution of the `Ansible` playbooks we launch `ssh-agent`
from the `setup_ssh` role itself. We register the environment variables
in the `Ansible` state and make use of them in the
`clone_personal_project` and `clone_work_projects` roles.

We also add the private keys to `ssh-agent` in the `setup_ssh` agent.

Containers are no longer using an entrypoint script. We execute `forger`
directly from the `podman run` call. This also help us having minimal
images compared to the base OS images we use.

Those changes were made with the intent of beter modularization of the
`Ansible` roles and easy maintanance in the future. It should also help
with adding support for other operating systems.