fix: Always export `ssh-agent` variables
We always create the same `~/.ssh/ssh-agent.socket` socket when we start
`ssh-agent` because of this we can always prepare the correct value for
`SSH_AUTH_SOCK`. For `SSH_AGENT_PID` we just use `pgrep` to find the PID
of `ssh-agent`.

This helps because even after a fresh run of the playbook when `fish`
starts it will have the `ssh-agent` variables set, even `ssh-agent` was
started by the playbook.

NOTE: The only thing to note is that the playbook needs to create the
same `~/.ssh/ssh-agent.socket` socket when starting `ssh-agent`.k