test: Add bulk testing on all supported environments
Modified `test_image.sh` to accept a `-it` optional parameter that will
dictate if we will get an interactive shell in the container after
Ansible did its job.

In adition we capture the exit code of `podman run` to be able to report
if the execution was successful or not.

Introduced a `test_all.sh` script that will launch a background job for
each supported environment. It basically runs `./test_image.sh <OS>`.

The `test_all.sh` script will help in testing on all environments after
we implemented a feature or fixed a bug on Fedora, which is my main
target for `forger`.