feat(setup_shell): Load `environment` file for `bash`
This is the first part of the work needed to be done to support a common
palce for declaring environment variables that will be sourced by both
`bash` and `fish`.

The location of the environment file is: `$HOME/.config/environment`.

The second part will cover loading the environment file in a `fish`
shell.

I don't know if I need to explain it again, but we do this because there
are times when I use `bash` in a terminal, also I want `bash` scripts to
have access to the environment variables I define.