feat(setup_shell): Load `environment` file for `fish`
This is the second part of `environment` loading. As the title suggests
we do it for `fish` this time.

The behavior is the same as for `bash`.

Now we can have the same environment for both `fish` and `bash`.

I've decided to do this because `fish` is not POSIX compliant. It's
scripting is nicer, but until I'm sure I can go full `fish` with this
one I'll do the best of both worlds: `fish` in terminal and `bash` in
services and other automation stuff.