feat(setup_shell): Merged `setup_eza` into `setup_shell`
I took the decision to merge `setup_eza` in `setup_shell` because in the
end I find it strange to setup a nice shell setup, but not use `eza` as
a replacement for `ls`.

The role `setup_shell` should setup a bare bones setup. I would go so
far to say that this is the minimum setup you have to do over Linux From
Scratch to have a bare bones and functional system. Most likely
`setup_shell` will be the only thing needed for user friendly container
too.

Maybe we'll go that route. We'll see, but for now I'm happy that this
reduce a bit the complexity of setting up a bare bones system.