fix(scripts): Use the `$SHELL` defined in the container
Because I did not escape `$` we ended executing the `$SHELL` defined on
the host.

Now escaping `$` we run the actual shell defined in the container.

The problem surfaced when I tried to apply `setup_ssh` in isolation and
`test_image.sh` was complaining it can't find `fish` inside the
containers. At first I tought I forgot a hardcoded `fish` path
somewhere, but no. It went undetected because until now I was testing
all roles or I always was applying `setup_shell` first. Once
`setup_shell` was applied then `fish` was present both on the host and
in the container.