refactor: Better secrets management
All the needed secrets from now on will be stored as encrypted vaults
inside the project. This commit introduces the SSH keys passphrase
stoared as an encrypted variable set in the `setup_ssh` role.

The container images will run the playbook and at the end they will
automatically. No more interactive mode. I'm looking forward to bring it
back, but for the time being I removed it.