chore(TODO): Remove the `TODO` file as it adds more overhead
I decided to turn to GitHub issues to track the stuff I have to do for
the project.

The `TODO` is problematic because different branches can have a
different state for the `TODO` file. Unifing the tracking of stuff to do
in the GitHub issues centralizes those items and it's easy to check what
is todo, what's in progress, if something was tried before, but
abandoned etc.