feat(task.shell): Backup `.bashrc` when deploying `forger`
We keep a history of the last 3 backups. Why?

There might be cases when we'll to a quick config change without
promoting it to `forger`. In this case we want to keep the old `.bashrc`
file just in case.

Haivng a backup will help us reapply the config, but also if we keep
facing the need to extract and reply something manually from the backup
it will be a good argument to promote them into the setup done by
`forger`.