fix(files.fish): Use `uname -n` to obtain the hostname
When `hostname` tool is not available our custom `fish` prompt will
fail.

Here is the exact error message:

```
Welcome to fish, the friendly interactive shell
Type help for instructions on how to use fish
fish: Unknown command: hostname
~/.config/fish/functions/fish_prompt.fish (line 1):
hostname
^~~~~~~^
in command substitution
	called on line 16 of file ~/.config/fish/functions/fish_prompt.fish
in function 'fish_prompt'
in command substitution
~/.config/fish/functions/fish_prompt.fish (line 16): Unknown command
    echo -n (hostname)
            ^~~~~~~~~^
in function 'fish_prompt'
in command substitution
[18:06] root@:.forger |
```

It's ugly to be meet with this.

To solve this we could install `hostname` by ourselves, but obtaining
the hostname using `uname -n` is much easier.

For the time being we'll adopt the `uname` solution and we'll adapt
again if this turns out to not be enough.