feat(files.fish): Display system uptime in greeter
It's good to remind ourselves when we open a new shell for how long the
system was up and running.

For me this works as a good reminder to check for updates or just do
routine restarts of the system.

The uptime is computed based on `/proc/uptime`.

There is no dependency on tools like `uptime`. The only dependency is
`awk`, but that can be eliminated if I really dedicate my neurons to it.

For math computations we use the built in `math` function from `fish`.