chore(files): Add template for the environment file
This is a environment file to define variables that will be available in
both `bash` and `fish` shells. This is it's only purpose.

It must use syntax supported by both `bash` and `fish`.

It's not entirely the same thing, but if it helps you to understand what
it does, it's the same thing like the environment files used to
initialize a container's environment variables, only that we use it for
host machine shells.