refactor(task.shell): Use Ansible built in backup functionality
Instead of rolling our own solution to backup the files we replace I
just figured out it's simple and more safe to use the built in
functionality from Ansible `copy` module. All I have to do is to set
`backup: true` and that's it!

This refactor also simplifies maintenance a lot. Sometimes reinventing
the wheel is worth it, but in this case I don't want to learn how to do
a backup scheme for configuration deployment. I actually want to get a
system up and running.