fix(files.env): Set `TERM` to `xterm-256color`
By default in most cases `TERM` is set to simple `xterm`. In that case
full color support in the terminal is not available. Because of this
`tmux` was not respecting the `brbright` color set by using `fish`
function `set_color`.

By setting `TERM=xterm-256color` we make sure the full color range is
supported.