feat(files.tmux): Launch `tmux-sm` using `Ctrl+a, e`
It is not enough to bind `Ctrl+e` at `fish` shell level. when we are
inside `tmux` we want to launch `tmux-sm` a bit differently.

When inside a `tmux` session a new window will be opened. In this window
the `tmux-sm` command will be executed. When a directory is chosen and
`tmux` switches to anohter session the window will close automatically.