refactor: Copy only the required files in each task group
It's easy to just copy `.config/fish/` for example, but the directory
may contain files from other task groups. In the case of `fish` configs
we have the `./config/conf.d/ssh-agent.fish` directory which is not need
for the `shell` tasks. It is needed for `ssh` tasks.

Copying only what we need lets us control exactly what we copy with a
group of tasks. The downside is that now we need to make sure the
destination directories exist and to have multipe copy commands, one per
needed file. Overall I think it's worth paying this price.

Also the file `.config/fish/conf.d/ssh-agent.fish` was renamed to
`.config/fish/conf.d/ssh_agent.fish` to have a consistent name schema.