fix(task.typewriter): Always force checkout `typewriter`
Until I can figure out how to deal with this in a better way I decided
to always hard reset `typewriter` clone when we run the task.

Keep in mind that we do a back up of the `nvim` directory, always
keeping the previous version around. Doing a diretory copy is not ideal
in my view, but it protects potential local changes.