fix(task.typewriter): Fix `$HOME/.config/nvim` backup
There were two issues caused by typos.

First typo `./config/.nvim` (BAD) vs `./config/nvim` (GOOD), when
checking for an old Neovim configuration. This typo lead to the old
Neovim configuration to not be saved in a backup directory.

Second typo `./config/nvim` (BAD) vs `./config/nvim/` (GOOD), when doing
the copy. The thing is that the version without `/` after `nvim` is not
bad as in "it's not working", it is bad because it will copy the
directory itself inside the `nvim.bak` directory. If you end the source
directory with a `/`, then the contents of the source directory will be
copied recusively in `nvim.bak`.

For more details check:
https://docs.ansible.com/ansible/latest/collections/ansible/builtin/copy_module.html#parameter-src