fix(files.tmux): Set terminal to `tmux-256color`
For a reason I don't really  understand it seems that using the terminal
type of  `screen-256color` not play nice with the color scheme. I tested
this inside _tmux_ running Neovim with the _gruvbox_ theme.

After setting `tmux-256color` as the terminal time the colors started to
render correctly in Neovim.