test(fedora_40): Remove because we focus QA only on Fedora 41
Fedora is not a rolling release, but it's a distribution that releases
often stable releases. In addition to this packages are not bleeding
edge, but they tend to follow upstream pretty close.

Because Fedora has no LTS releases the target is to be compatible with
the latest stable release, Fedora 41 at the moment of writing.

When using Fedora keeping up with the latest stable release it's really
easy with the CLI upgrader so there's no reason to lag behind. Maybe at
some point `foreger` will also do system upgrades? That's another story,
but for now lets focus only on the latest stable release.